add POC server iniated client.
The snap agent initiates a websocket connection to a server.
We then use this websocket as a transport for the remote server
to send REST calls.

To minimize code changes to Snap, we are encapsulating the REST
client HTTP calls inside JSONRPC calls over  websocket, which
is over a HTTP interface.  With this approach the "client", which
is the websocket server, can use the native snap client library
to manage the node.

There is definitely cleaner ways to implement all of this, but
this demonstrates the concept of a client with a server initiated
connection nicely.

A basic reference server is provide in mgmt/rest/ws_client/server